victimas_accionar_represivo_ilegal_sin_denuncia_formal

se detalla el listado de casos en investigación incluidos en el listado de la Conadep sin denuncia formal ante la Secretaría de Derechos Humanos y Pluralismo Cultural de la Nación, que revisten la categoría de “presunción de desaparición forzada”. Recurso original: http://datos.jus.gob.ar/dataset/registro-unificado-de-victimas-del-terrorismo-de-estado-ruvte

victimas_accionar_represivo_ilegal_sin_denuncia_formal

Format

A data frame with 784 rows and 21 variables:

id_unico_ruvte

character número de identificación del caso individual

anio_denuncia

double año de alta original de la denuncia en los registros que hoy constituyen el Ruvte

tipificacion_ruvte

character tipificación principal del caso según el Ruvte (víctima de desaparición forzada o víctima de asesinato), a la que se agregan observaciones que en algunos casos especifican la fuente, y en otros, actualizaciones de estado (por ejemplo, exhumados e identificados sus restos, entre otras)

apellido_paterno_nombres

character apellido paterno y nombres de pila de la víctima

apellido_materno

character apellido materno

apellido_casada

character apellido de casada

edad_al_momento_del_hecho

character edad al momento del hecho (al momento de su deceso, o, en su defecto, de su “desaparición”)

edad_al_momento_del_hecho_numerico

numérico edad al momento del hecho en años. 0 significa nacido en cautiverio.

documentos

character tipo y número de documento

anio_nacimiento

character año de nacimiento

provincia_nacimiento

character provincia de nacimiento (solo para el caso de ciudadanos nacidos en Argentina)

pais_nacimiento

character país de nacimiento

nacionalidad

character nacionalidad de la víctima

embarazo

logical datos sobre el embarazo de la víctima y sobre el destino del niño nacido en cautiverio

fecha_detencion_secuestro

character fecha de detención-secuestro de la víctima

lugar_detencion_secuestro

character lugar de detención-secuestro de la víctima

fecha_asesinato_o_hallazgo_de_restos

character fecha de asesinato, fallecimiento o hallazgo de los restos de la víctima

lugar_asesinato_o_hallazgo_de_restos

character lugar de asesinato, fallecimiento o hallazgo de los restos de la víctima

fotografia

character información sobre si obra fotografía de la víctima en base de datos

fuente_original

character fuente original de la denuncia, en base a la cual fue incluida en el listado de la Conadep de 1984

provincia_nacimiento_indec_id

character provincia de nacimiento (solo para el caso de ciudadanos nacidos en Argentina), según la codificación de provincia implementada por INDEC

pais_nacimiento_indec_id

character país de nacimiento, según la codificación de país implementada por INDEC.